[android] fix metadata conversion:
